Extract HiveClientPool as an abstract class to facilitate testing
